requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

cryptography
azure-functions
azure-functions-durable
azure-ai-documentintelligence
azure-storage-blob
azure-identity
azure-ai-formrecognizer
requests
pandas
numpy
PyPDF2
openai>=1.25.0
azure-functions
